+++ src/wfg/native_ui/ui/components/LayoutOffset.java
@@ -1,6 +1,6 @@
 package wfg.native_ui.ui.components;
 
-public final class LayoutOffsetComp extends BaseComponent {
+public final class LayoutOffset extends BaseComponent {
     public int x, y, w, h;
 
     public final void setOffset(int x, int y, int w, int h) {