+++ src/wfg/native_ui/ui/components/OutlineComp.java
@@ -7,6 +7,7 @@ import java.awt.Color;
 public final class OutlineComp extends BaseComponent {
     public OutlineType type = OutlineType.LINE;
     public Color color = dark;
+    public final LayoutOffset offset = new LayoutOffset();
 
     public static enum OutlineType {
         LINE,