+++ src/wfg/native_ui/ui/core/UIElementFlags.java
@@ -7,7 +7,6 @@ public interface UIElementFlags {
     public interface HasAudioFeedback {}
     public interface HasBackground {}
     public interface HasTooltip {}
-    public interface HasLayoutOffset {}
     public interface HasUIContext {}
     public interface HasInputSnapshot {}
 }